### 2018-07-27 #241 @MahShaaban
Pre-submisstion Inquiry: colocr
I developed this simple package `colocr` for conducting the co-localization analysis of microscopy images. The package comes with a shinyapp to interactively choose the regions of interest.  And it calculates two simple statistics to quantify the co-localization of two proteins in a certain cell.

I wonder whether this would be of interest to rOpenSci.

The GitHub repo is [https://github.com/MahShaaban/colocr](https://github.com/MahShaaban/colocr)
