### 2018-07-29 #241 @maelle
Thanks for your pre-submission inquiry @mahshaaban! Can you confirm that the primary goal of the package is to compute **statistics** from images? Methods packages are out of scope.

For reference https://ropensci.github.io/dev_guide/policies.html#aims-and-scope
