### 2018-07-30 #241 @MahShaaban
Thanks, @maelle. I am aware of the fact that methods packages are out of the scope. Although the main use of the package is to compute co-localization stats, this is not a significant part of what the package does. It's just a one call to `cor()`. What the package focuses on automating the selection of regions of interest in the images. In addition, the package contains a shiny app to interactively perform the same task.
