### 2018-07-31 #241 @maelle
Sorry to ask so many questions, but how do people usually get these statistics, would you argue that with this package they'd be more prone to using a script rather than a click-and-point software?
