### 2018-08-01 #241 @MahShaaban
That's alright.
I'd say yes. There are two cases where using scripts is more convenient. These are

1. Analyzing a large number of images
2. Maintaining a reproducible workflow

In my research, I care a lot about reproducibility. While I mainly use high throughput data, often the case my work is followed by wet lab experiments for validation. Analyzing 10s of pictures may be doable in programs like **ImageJ** but would make it difficult to use the raw images in the study workflow. This was the reason I started this simple package.
