### 2018-08-02 #241 @maelle
Thanks @MahShaaban!

After discussion we agree that your package is in scope so please submit it when it's ready!

A note about two rOpenSci's packages that **might** be useful
* [`magick`](https://github.com/ropensci/magick)
* [`ijtiff`](https://github.com/ropensci/ijtiff)
